refactor(frontend): relocate plugin consts
